+++ sdk/cosmosdb/cosmos/src/common/helper.ts
@@ -314,11 +314,3 @@ export function parseConnectionString(connectionString: string): CosmosClientOpt
     key: AccountKey
   };
 }
-
-/**
- * @ignore
- */
-// https://github.com/iliakan/detect-node/blob/master/index.js
-export const isNode: boolean =
-  Object.prototype.toString.call(typeof process !== "undefined" ? process : 0) ===
-  "[object process]";