+++ sdk/identity/identity/src/credentials/authorizationCodeCredential.ts
@@ -9,6 +9,7 @@ import { IdentityClient, TokenResponse, TokenCredentialOptions } from "../client
 import { CanonicalCode } from "@opentelemetry/api";
 import { credentialLogger, formatSuccess, formatError } from "../util/logging";
 import { getIdentityTokenEndpointSuffix } from "../util/identityTokenEndpoint";
+import { checkTenantId } from "../util/checkTenantId";
 
 const logger = credentialLogger("AuthorizationCodeCredential");
 
@@ -98,6 +99,8 @@ export class AuthorizationCodeCredential implements TokenCredential {
     redirectUriOrOptions: string | TokenCredentialOptions | undefined,
     options?: TokenCredentialOptions
   ) {
+    checkTenantId(logger, tenantId);
+
     this.clientId = clientId;
     this.tenantId = tenantId;