+++ sdk/metricsadvisor/ai-metrics-advisor/recordings/node/metricsadvisoradministrationclient_datafeed_datafeed/recording_creates_an_azure_cosmos_db_feed.js
@@ -1,47 +1,51 @@
 let nock = require('nock');
 
-module.exports.hash = "a8710b5da20f47e300e40bc429789016";
+module.exports.hash = "27fe79fb9a493fbeffc6a93e9a5f771a";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
 nock('https://endpoint:443', {"encodedQueryParams":true})
-  .post('/metricsadvisor/v1.0/dataFeeds', {"dataSourceType":"AzureCosmosDB","dataFeedName":"js-test-cosmosFeed-160072575346106475","granularityName":"Daily","metrics":[{"metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00.000Z","startOffsetInSeconds":0,"maxConcurrency":-1,"minRetryIntervalInSeconds":-1,"stopRetryAfterInSeconds":-1,"needRollup":"NeedRollup","rollUpMethod":"Sum","allUpIdentification":"__CUSTOM_SUM__","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"viewMode":"Private","dataSourceParameter":{"connectionString":"Server=server.example.net;Encrypt=True;","sqlQuery":"let starttime=datetime(@StartTime); let endtime=starttime","database":"sample","collectionId":"sample"}})
+  .post('/metricsadvisor/v1.0/dataFeeds', {"dataSourceType":"AzureCosmosDB","dataFeedName":"js-test-cosmosFeed-160323420510600639","dataFeedDescription":"Data feed description","granularityName":"Daily","metrics":[{"metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00.000Z","startOffsetInSeconds":0,"maxConcurrency":-1,"minRetryIntervalInSeconds":-1,"stopRetryAfterInSeconds":-1,"needRollup":"NeedRollup","rollUpMethod":"Sum","allUpIdentification":"__CUSTOM_SUM__","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"viewMode":"Private","dataSourceParameter":{"connectionString":"Server=server.example.net;Encrypt=True;","sqlQuery":"let starttime=datetime(@StartTime); let endtime=starttime","database":"sample","collectionId":"sample"}})
   .reply(201, "", [
   'Content-Length',
   '0',
   'Location',
-  'https://endpoint/metricsadvisor/v1.0/dataFeeds/d32b4a7c-b8f4-4868-9a07-5e885d5840d5',
+  'https://endpoint/metricsadvisor/v1.0/dataFeeds/5b7e9583-8bc5-4af3-88d2-367497dc5f03',
   'x-request-id',
-  '300ca7c6-be19-42ed-83c6-a8bc068a7c22',
+  '3b311eea-43b2-4090-acd5-3e828d939956',
   'x-envoy-upstream-service-time',
-  '355',
+  '2305',
   'apim-request-id',
-  '300ca7c6-be19-42ed-83c6-a8bc068a7c22',
+  '3b311eea-43b2-4090-acd5-3e828d939956',
   'Strict-Transport-Security',
   'max-age=31536000; includeSubDomains; preload',
   'x-content-type-options',
   'nosniff',
   'Date',
-  'Mon, 21 Sep 2020 22:02:44 GMT'
+  'Tue, 20 Oct 2020 22:50:41 GMT',
+  'Connection',
+  'close'
 ]);
 
 nock('https://endpoint:443', {"encodedQueryParams":true})
-  .get('/metricsadvisor/v1.0/dataFeeds/d32b4a7c-b8f4-4868-9a07-5e885d5840d5')
-  .reply(200, {"dataFeedId":"d32b4a7c-b8f4-4868-9a07-5e885d5840d5","dataFeedName":"js-test-cosmosFeed-160072575346106475","metrics":[{"metricId":"d6519a2f-03fc-45ed-b71c-7e9cc788f295","metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricId":"4e325640-bcc5-4f82-a269-7a3b60b2ff1b","metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00Z","dataSourceType":"AzureCosmosDB","timestampColumn":"","startOffsetInSeconds":0,"maxQueryPerMinute":30,"granularityName":"Daily","granularityAmount":null,"allUpIdentification":"__CUSTOM_SUM__","needRollup":"NeedRollup","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"rollUpMethod":"Sum","rollUpColumns":[],"dataFeedDescription":"","stopRetryAfterInSeconds":-1,"minRetryIntervalInSeconds":-1,"maxConcurrency":-1,"viewMode":"Private","admins":["yumeng@microsoft.com"],"viewers":[],"creator":"yumeng@microsoft.com","status":"Active","createdTime":"2020-09-21T22:02:44Z","isAdmin":true,"actionLinkTemplate":"","dataSourceParameter":{"connectionString":"Server=server.example.net;Encrypt=True;","database":"sample","sqlQuery":"let starttime=datetime(@StartTime); let endtime=starttime","collectionId":"sample"}}, [
+  .get('/metricsadvisor/v1.0/dataFeeds/5b7e9583-8bc5-4af3-88d2-367497dc5f03')
+  .reply(200, {"dataFeedId":"5b7e9583-8bc5-4af3-88d2-367497dc5f03","dataFeedName":"js-test-cosmosFeed-160323420510600639","metrics":[{"metricId":"b61f8fe3-5c3a-488b-8dd1-c309211f8333","metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricId":"339421ba-8eb7-4960-8b1b-062ebe2bd82b","metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00Z","dataSourceType":"AzureCosmosDB","timestampColumn":"","startOffsetInSeconds":0,"maxQueryPerMinute":30,"granularityName":"Daily","granularityAmount":null,"allUpIdentification":"__CUSTOM_SUM__","needRollup":"NeedRollup","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"rollUpMethod":"Sum","rollUpColumns":[],"dataFeedDescription":"Data feed description","stopRetryAfterInSeconds":-1,"minRetryIntervalInSeconds":-1,"maxConcurrency":-1,"viewMode":"Private","admins":["yumeng@microsoft.com"],"viewers":[],"creator":"yumeng@microsoft.com","status":"Active","createdTime":"2020-10-20T22:50:42Z","isAdmin":true,"actionLinkTemplate":"","dataSourceParameter":{"connectionString":"Server=server.example.net;Encrypt=True;","database":"sample","sqlQuery":"let starttime=datetime(@StartTime); let endtime=starttime","collectionId":"sample"}}, [
   'Content-Length',
-  '1382',
+  '1403',
   'Content-Type',
   'application/json; charset=utf-8',
   'x-request-id',
-  '4aa5a030-fc70-4cc7-9403-66991aabadf5',
+  'cf1164da-a14e-48d9-9e6f-dd97e4998f37',
   'x-envoy-upstream-service-time',
-  '134',
+  '143',
   'apim-request-id',
-  '4aa5a030-fc70-4cc7-9403-66991aabadf5',
+  'cf1164da-a14e-48d9-9e6f-dd97e4998f37',
   'Strict-Transport-Security',
   'max-age=31536000; includeSubDomains; preload',
   'x-content-type-options',
   'nosniff',
   'Date',
-  'Mon, 21 Sep 2020 22:02:44 GMT'
+  'Tue, 20 Oct 2020 22:50:42 GMT',
+  'Connection',
+  'close'
 ]);