+++ sdk/metricsadvisor/ai-metrics-advisor/recordings/node/metricsadvisoradministrationclient_datafeed_datafeed/recording_creates_mongodb_data_feed.js
@@ -1,47 +1,51 @@
 let nock = require('nock');
 
-module.exports.hash = "30517847d1a476644067e3b00136c324";
+module.exports.hash = "e3fb495d8a0717ebae138ef177872e0c";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
 nock('https://endpoint:443', {"encodedQueryParams":true})
-  .post('/metricsadvisor/v1.0/dataFeeds', {"dataSourceType":"MongoDB","dataFeedName":"js-test-mongoDbFeed-160072575346109637","granularityName":"Daily","metrics":[{"metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00.000Z","startOffsetInSeconds":0,"maxConcurrency":-1,"minRetryIntervalInSeconds":-1,"stopRetryAfterInSeconds":-1,"needRollup":"NeedRollup","rollUpMethod":"Sum","allUpIdentification":"__CUSTOM_SUM__","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"viewMode":"Private","dataSourceParameter":{"connectionString":"https://connect-to-mongodb","database":"data-feed-mongodb","command":"{ find: mongodb,filter: { Time: @StartTime },batch: 200 }"}})
+  .post('/metricsadvisor/v1.0/dataFeeds', {"dataSourceType":"MongoDB","dataFeedName":"js-test-mongoDbFeed-160323420510601647","dataFeedDescription":"Data feed description","granularityName":"Daily","metrics":[{"metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00.000Z","startOffsetInSeconds":0,"maxConcurrency":-1,"minRetryIntervalInSeconds":-1,"stopRetryAfterInSeconds":-1,"needRollup":"NeedRollup","rollUpMethod":"Sum","allUpIdentification":"__CUSTOM_SUM__","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"viewMode":"Private","dataSourceParameter":{"connectionString":"https://connect-to-mongodb","database":"data-feed-mongodb","command":"{ find: mongodb,filter: { Time: @StartTime },batch: 200 }"}})
   .reply(201, "", [
   'Content-Length',
   '0',
   'Location',
-  'https://endpoint/metricsadvisor/v1.0/dataFeeds/45c51238-b2bf-46a3-ae52-148ba9d45330',
+  'https://endpoint/metricsadvisor/v1.0/dataFeeds/2ff20eac-b64c-4bde-95d4-8feba3b2d311',
   'x-request-id',
-  '7a281151-94e4-4106-9c67-8a49e4c29040',
+  '2ce5a38c-df9e-472f-8cb3-f7fef3c34d62',
   'x-envoy-upstream-service-time',
-  '397',
+  '281',
   'apim-request-id',
-  '7a281151-94e4-4106-9c67-8a49e4c29040',
+  '2ce5a38c-df9e-472f-8cb3-f7fef3c34d62',
   'Strict-Transport-Security',
   'max-age=31536000; includeSubDomains; preload',
   'x-content-type-options',
   'nosniff',
   'Date',
-  'Mon, 21 Sep 2020 22:02:57 GMT'
+  'Tue, 20 Oct 2020 22:50:48 GMT',
+  'Connection',
+  'close'
 ]);
 
 nock('https://endpoint:443', {"encodedQueryParams":true})
-  .get('/metricsadvisor/v1.0/dataFeeds/45c51238-b2bf-46a3-ae52-148ba9d45330')
-  .reply(200, {"dataFeedId":"45c51238-b2bf-46a3-ae52-148ba9d45330","dataFeedName":"js-test-mongoDbFeed-160072575346109637","metrics":[{"metricId":"911927bb-2358-408e-91be-504559f73952","metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricId":"027f1ca2-a1e0-49bc-ab89-1648d80ade2e","metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00Z","dataSourceType":"MongoDB","timestampColumn":"","startOffsetInSeconds":0,"maxQueryPerMinute":30,"granularityName":"Daily","granularityAmount":null,"allUpIdentification":"__CUSTOM_SUM__","needRollup":"NeedRollup","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"rollUpMethod":"Sum","rollUpColumns":[],"dataFeedDescription":"","stopRetryAfterInSeconds":-1,"minRetryIntervalInSeconds":-1,"maxConcurrency":-1,"viewMode":"Private","admins":["yumeng@microsoft.com"],"viewers":[],"creator":"yumeng@microsoft.com","status":"Active","createdTime":"2020-09-21T22:02:57Z","isAdmin":true,"actionLinkTemplate":"","dataSourceParameter":{"connectionString":"https://connect-to-mongodb","database":"data-feed-mongodb","command":"{ find: mongodb,filter: { Time: @StartTime },batch: 200 }"}}, [
+  .get('/metricsadvisor/v1.0/dataFeeds/2ff20eac-b64c-4bde-95d4-8feba3b2d311')
+  .reply(200, {"dataFeedId":"2ff20eac-b64c-4bde-95d4-8feba3b2d311","dataFeedName":"js-test-mongoDbFeed-160323420510601647","metrics":[{"metricId":"64f64c4e-56b7-4c90-af02-a74ad09d9746","metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricId":"d57d4b19-9356-4739-994f-f8e33971b8b8","metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00Z","dataSourceType":"MongoDB","timestampColumn":"","startOffsetInSeconds":0,"maxQueryPerMinute":30,"granularityName":"Daily","granularityAmount":null,"allUpIdentification":"__CUSTOM_SUM__","needRollup":"NeedRollup","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"rollUpMethod":"Sum","rollUpColumns":[],"dataFeedDescription":"Data feed description","stopRetryAfterInSeconds":-1,"minRetryIntervalInSeconds":-1,"maxConcurrency":-1,"viewMode":"Private","admins":["yumeng@microsoft.com"],"viewers":[],"creator":"yumeng@microsoft.com","status":"Active","createdTime":"2020-10-20T22:50:48Z","isAdmin":true,"actionLinkTemplate":"","dataSourceParameter":{"connectionString":"https://connect-to-mongodb","database":"data-feed-mongodb","command":"{ find: mongodb,filter: { Time: @StartTime },batch: 200 }"}}, [
   'Content-Length',
-  '1350',
+  '1371',
   'Content-Type',
   'application/json; charset=utf-8',
   'x-request-id',
-  'bba7d5fe-b701-4fe9-b1d9-dab4b1ee4f65',
+  '699262e0-ff55-438a-8530-0d94aec08ce7',
   'x-envoy-upstream-service-time',
-  '217',
+  '109',
   'apim-request-id',
-  'bba7d5fe-b701-4fe9-b1d9-dab4b1ee4f65',
+  '699262e0-ff55-438a-8530-0d94aec08ce7',
   'Strict-Transport-Security',
   'max-age=31536000; includeSubDomains; preload',
   'x-content-type-options',
   'nosniff',
   'Date',
-  'Mon, 21 Sep 2020 22:02:57 GMT'
+  'Tue, 20 Oct 2020 22:50:48 GMT',
+  'Connection',
+  'close'
 ]);