+++ sdk/metricsadvisor/ai-metrics-advisor/samples/javascript/incidentsAndAlerts.js
@@ -37,10 +37,10 @@ async function listIncidentsForDetectionConfig(client, detectionConfigId) {
   console.log("  using for-await-of syntax");
   for await (const incident of client.listIncidentsForDetectionConfiguration(
     detectionConfigId,
-    new Date("09/06/2020"),
-    new Date("09/11/2020"),
+    new Date("10/22/2020"),
+    new Date("10/24/2020"),
     {
-      dimensionFilter: [{ dimension: { city: "Manila", category: "Shoes Handbags & Sunglasses" } }]
+      dimensionFilter: [{ city: "Manila", category: "Shoes Handbags & Sunglasses" }]
     }
   )) {
     console.log("    Incident");
@@ -49,7 +49,7 @@ async function listIncidentsForDetectionConfig(client, detectionConfigId) {
     console.log(`      status: ${incident.status}`);
     console.log(`      startTime: ${incident.rootDimensionKey}`);
     console.log(`      startTime: ${incident.startTime}`);
-    console.log(`      last occured: ${incident.lastOccuredTime}`);
+    console.log(`      last occured: ${incident.lastOccurredTime}`);
     console.log(`      detection config id: ${incident.detectionConfigurationId}`);
   }
 
@@ -57,8 +57,8 @@ async function listIncidentsForDetectionConfig(client, detectionConfigId) {
   const iterator = client
     .listIncidentsForDetectionConfiguration(
       detectionConfigId,
-      new Date("09/06/2020"),
-      new Date("09/11/2020")
+      new Date("10/22/2020"),
+      new Date("10/24/2020")
     )
     .byPage({ maxPageSize: 20 });
   let result = await iterator.next();
@@ -71,7 +71,7 @@ async function listIncidentsForDetectionConfig(client, detectionConfigId) {
       "status",
       "rootDimensionKey",
       "startTime",
-      "lastOccuredTime",
+      "lastOccurredTime",
       "detectionConfigurationId"
     ]);
     result = await iterator.next();
@@ -83,8 +83,8 @@ async function listAnomaliesForDetectionConfig(client, detectionConfigId) {
   console.log("  using for-await-of syntax");
   for await (const anomaly of client.listAnomaliesForDetectionConfiguration(
     detectionConfigId,
-    new Date("09/06/2020"),
-    new Date("09/11/2020"),
+    new Date("10/22/2020"),
+    new Date("10/24/2020"),
     {
       severityFilter: { min: "Medium", max: "High" }
     }
@@ -98,13 +98,12 @@ async function listAnomaliesForDetectionConfig(client, detectionConfigId) {
     console.log(`      status: ${anomaly.status}`);
     console.log(`      series key: ${anomaly.seriesKey}`);
   }
-
   console.log(`  by pages`);
   const iterator = client
     .listAnomaliesForDetectionConfiguration(
       detectionConfigId,
-      new Date("09/06/2020"),
-      new Date("09/11/2020"),
+      new Date("10/22/2020"),
+      new Date("10/24/2020"),
       {
         severityFilter: { min: "Medium", max: "High" }
       }
@@ -136,8 +135,8 @@ async function listAlerts(client, alertConfigId) {
   console.log("  using for-await-of syntax");
   for await (const alert of client.listAlertsForAlertConfiguration(
     alertConfigId,
-    new Date("01/01/2020"),
-    new Date("09/09/2020"),
+    new Date("10/22/2020"),
+    new Date("10/24/2020"),
     "AnomalyTime"
   )) {
     console.log("    Alert");
@@ -150,8 +149,8 @@ async function listAlerts(client, alertConfigId) {
   const iterator = client
     .listAlertsForAlertConfiguration(
       alertConfigId,
-      new Date("01/01/2020"),
-      new Date("09/09/2020"),
+      new Date("10/22/2020"),
+      new Date("10/24/2020"),
       "AnomalyTime"
     )
     .byPage({ maxPageSize: 20 });
@@ -176,7 +175,7 @@ async function listIncidentsForAlert(client, alertConfigId, alertId) {
     console.log(`      status: ${incident.status}`);
     console.log(`      startTime: ${incident.rootDimensionKey}`);
     console.log(`      startTime: ${incident.startTime}`);
-    console.log(`      last occured: ${incident.lastOccuredTime}`);
+    console.log(`      last occured: ${incident.lastOccurredTime}`);
     console.log(`      detection config id: ${incident.detectionConfigurationId}`);
   }
 
@@ -192,7 +191,7 @@ async function listIncidentsForAlert(client, alertConfigId, alertId) {
       "status",
       "rootDimensionKey",
       "startTime",
-      "lastOccuredTime",
+      "lastOccurredTime",
       "detectionConfigurationId"
     ]);
     result = await iterator.next();
@@ -214,7 +213,6 @@ async function listAnomaliesForAlert(client, alertConfigId, alertId) {
     console.log(`      status: ${anomaly.status}`);
     console.log(`      series key: ${anomaly.seriesKey}`);
   }
-
   console.log(`  by pages`);
   const iterator = client.listAnomaliesForAlert(alertConfigId, alertId).byPage({ maxPageSize: 20 });