+++ sdk/metricsadvisor/ai-metrics-advisor/samples/typescript/src/seriesData.ts
@@ -37,8 +37,8 @@ async function getEnrichedSeriesData(client: MetricsAdvisorClient, detectionConf
       new Date("09/01/2020"),
       new Date("09/12/2020"),
       [
-        { dimension: { Dim1: "Common Lime", Dim2: "Amphibian" } },
-        { dimension: { Dim1: "Common Beech", Dim2: "Ant" } }
+        { city: "Manila", category: "Handmade" },
+        { city: "Shanghai", category: "Shoes Handbags & Sunglasses" }
       ]
     );
 
@@ -72,8 +72,8 @@ async function getMetricSeriesData(client: MetricsAdvisorClient, metricId: strin
       new Date("09/01/2020"),
       new Date("09/12/2020"),
       [
-        { dimension: { Dim1: "Common Lime", Dim2: "Amphibian" } },
-        { dimension: { Dim1: "Common Beech", Dim2: "Ant" } }
+        { city: "Manila", category: "Handmade" },
+        { city: "Shanghai", category: "Shoes Handbags & Sunglasses" }
       ]
     );