+++ sdk/servicebus/service-bus/src/index.ts
@@ -59,9 +59,6 @@ export {
 } from "./serviceBusAtomManagementClient";
 export { ServiceBusClient } from "./serviceBusClient";
 export {
-  AmqpAnnotatedMessage,
-  AmqpMessageHeader,
-  AmqpMessageProperties,
   DeadLetterOptions,
   ServiceBusReceivedMessage,
   ServiceBusMessage